Reads TR username and password from keys.property file and adds them base64 encoded to TierekisteriClient request headers
