+++ digiroad2-oracle/src/test/resources/keys.properties
@@ -1,2 +1,4 @@
 googlemapapi.client_id=XYZ123
 googlemapapi.crypto_key=ZYX321
+tierekisteri.username=userXYZ
+tierekisteri.password=passwordXYZ
